UK2GTFS: Convert UK Transport Timetable Datasets to GTFS Format

Convert transport timetable data from the formats used in the United Kingdom to the General Transit Feed Specification ('GTFS') format. Supports the Common Interface File ('CIF') format used by the Rail Delivery Group and Network Rail for heavy and light rail timetables, 'TransXChange' XML files used for bus, coach, and ferry timetables, and the historic National Public Transport Data Repository ('NPTDR') archives. Also provides utilities for reading, writing, cleaning, subsetting, and merging 'GTFS' files.

See also

Author

Maintainer: Malcolm Morgan m.morgan1@leeds.ac.uk (ORCID) [copyright holder]

Authors:

Owen O'Neill owen@wvr.org.uk (ORCID)

Other contributors:

Adrian Schönig [contributor]