ATOC to GTFS

Convert ATOC CIF files to GTFS

atoc2gtfs(
  path_in,
  silent = TRUE,
  ncores = 1,
  locations = tiplocs,
  agency = atoc_agency,
  shapes = FALSE,
  transfers = TRUE
)

Arguments

path_in

Character, path to ATOC file e.g."C:/input/ttis123.zip"

silent

Logical, should progress messages be surpressed (default TRUE)

ncores

Numeric, When parallel processing how many cores to use (default 1)

locations

where to get tiploc locations (see details)

agency

where to get agency.txt (see details)

shapes

Logical, should shapes.txt be generated (default FALSE)

transfers

Logical, should transfers.txt be generated (default TRUE)

Details

Locations

The .msn file contains the physical locations of stations and other TIPLOC codes (e.g. junctions). However, the quality of the locations is often poor only accurate to about 1km and occasionally very wrong. Therefore, the UK2GTFS package contains an internal dataset of the TIPLOC locations with better location accuracy, which are used by default.

However you can also specify `locations = "file"` to use the TIPLOC locations in the ATOC data or provide an SF data frame of your own.

Agency

The ATOC files do not contain the necessary information to build the agency.txt file. Therfore this data is provided with the package. You can also pass your own data frame of agency information.

See also