TransXchange to GTFS

TransXchange to GTFS

transxchange2gtfs(
  path_in,
  silent = TRUE,
  ncores = 1,
  cal = get_bank_holidays(),
  naptan = get_naptan(),
  scotland = "auto",
  try_mode = TRUE,
  force_merge = FALSE
)

Arguments

path_in

Path to zipped transxchange files

silent

Logical, should progress be shown

ncores

Numeric, When parallel processing how many cores to use

cal

Calendar object from get_bank_holidays()

naptan

Naptan stop locations from get_naptan()

scotland

character, should Scottish bank holidays be used? Can be "auto" (defualt), "yes", "no". If "auto" and path_in ends with "S.zip" Scottish bank holidays will be used, otherwise England and Wales bank holidays are used.

try_mode

Logical, if TRUE import and conversion are wrapped in try calls thus a failure on a single file will not cause the whole process to fail. Warning this could result in a GTFS file with missing routes.

force_merge

Logical, passed to gtfs_merge(force), default FALSE

Value

A GTFS named list

Details

Convert transxchange files to GTFS

This is a meta fucntion which aids TransXchange to GTFS conversion. It simple runs transxchange_import(), transxchange_export(), gtfs_merge(), gtfs_write()

Progress Bars

To minimise overall processing when using mulitple cores the fucntion works from largest to smallest file.This can mean the progress bar sits a 0 quite some time, before starting to move rapidly.